Extract shared MCP app helpers
Move the FastMCP app instance plus `_result` and `_error_result` into
`server_app.py`, while keeping `server.py` as the public registration and
re-export surface. This is the first server decomposition step and keeps all
registered tools, resources, and direct imports intact.

Constraint: Preserve `from mcp_video.server import mcp, _result, _error_result` compatibility
Rejected: Move tool registrations in the same change | shared app extraction should land before registration modules split
Confidence: high
Scope-risk: narrow
Tested: python3 -m pytest tests/test_server.py::TestServerInitialization tests/test_server.py::TestErrorHandling tests/test_public_surface.py -q --tb=short; python3 -m ruff check mcp_video/server.py mcp_video/server_app.py tests/test_server.py tests/test_public_surface.py; python3 -m ruff format --check mcp_video/server.py mcp_video/server_app.py; async smoke of mcp.list_tools() returning 83 tools
Not-tested: full non-slow suite after server app extraction
